tp3 - Ordenacao ok, falta mediana e lim_m e organizar

+++ tps/tp3/bib-qse/bib-qse-copia-testes/area.h
@@ -12,7 +12,7 @@
 typedef int TipoApontador;
 
 typedef struct TipoRegistro {
-  long long int Chave;
+  float Chave;
 } TipoRegistro;
 
 typedef struct TipoCelula {
